refactor: use failsafe for engine/exporter tests
