+++ engine/src/test/java/io/zeebe/containers/engine/ZeebeClusterEngineIT.java
@@ -22,8 +22,9 @@ import static org.assertj.core.api.Assertions.within;
 import io.camunda.zeebe.client.ZeebeClient;
 import io.camunda.zeebe.client.api.command.FinalCommandStep;
 import io.camunda.zeebe.protocol.record.Record;
-import io.zeebe.containers.ZeebeContainer;
+import io.zeebe.containers.ZeebeGatewayNode;
 import io.zeebe.containers.clock.ZeebeClock;
+import io.zeebe.containers.cluster.ZeebeCluster;
 import io.zeebe.containers.exporter.DebugReceiver;
 import java.io.IOException;
 import java.net.ConnectException;
@@ -32,28 +33,46 @@ import java.net.Socket;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.Future;
 import java.util.concurrent.RejectedExecutionException;
+import java.util.stream.Collectors;
+import org.agrona.CloseHelper;
 import org.awaitility.Awaitility;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.testcontainers.containers.Network;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers
-final class ZeebeContainerEngineTest {
+final class ZeebeClusterEngineIT {
+  private final Network network = Network.newNetwork();
   private final InfiniteList<Record<?>> records = new InfiniteList<>();
   private final DebugReceiver receiver = new DebugReceiver(records::add);
   private final DebugReceiverStream recordStream = new DebugReceiverStream(records, receiver);
-  private final ZeebeContainer container = new ZeebeContainer();
+  private final ZeebeCluster cluster =
+      ZeebeCluster.builder()
+          .withEmbeddedGateway(true)
+          .withBrokersCount(2)
+          .withPartitionsCount(1)
+          .withReplicationFactor(2)
+          .build();
+
+  @AfterEach
+  void afterEach() {
+    CloseHelper.close(network);
+  }
 
   @Test
   void shouldCloseEverythingOnStop() {
     // given
     final ZeebeClient client;
     final InetSocketAddress receiverAddress;
-    try (final ZeebeContainerEngine<?> engine =
-        new ZeebeContainerEngine<>(container, recordStream)) {
+    try (final ZeebeClusterEngine engine = new ZeebeClusterEngine(cluster, recordStream)) {
       engine.start();
       client = engine.createClient();
       receiverAddress = receiver.serverAddress();
@@ -68,7 +87,7 @@ final class ZeebeContainerEngineTest {
               assertThatCode(request::send)
                   .hasRootCauseInstanceOf(RejectedExecutionException.class);
             });
-    assertThat(container.isStarted()).isFalse();
+    assertThat(cluster.getNodes().values()).allMatch(node -> !node.isRunning());
     assertThatCode(() -> testServerConnection(receiverAddress))
         .isInstanceOf(ConnectException.class);
   }
@@ -80,10 +99,8 @@ final class ZeebeContainerEngineTest {
   }
 
   @Nested
-  final class WithContainerTest {
-    @Container
-    private final ZeebeContainerEngine<?> engine =
-        new ZeebeContainerEngine<>(container, recordStream);
+  final class WithClusterTest {
+    @Container final ZeebeClusterEngine engine = new ZeebeClusterEngine(cluster, recordStream);
 
     @Test
     void shouldCreateClient() {
@@ -105,22 +122,34 @@ final class ZeebeContainerEngineTest {
       final String address = engine.getGatewayAddress();
 
       // then
-      assertThat(address).isEqualTo(container.getExternalGatewayAddress());
+      final Set<String> gatewayAddress =
+          cluster.getGateways().values().stream()
+              .map(ZeebeGatewayNode::getExternalGatewayAddress)
+              .collect(Collectors.toSet());
+      assertThat(address).isIn(gatewayAddress);
     }
 
     @Test
     void shouldIncreaseTime() {
       // given
       final Duration offset = Duration.ofMinutes(5);
-      final ZeebeClock clock = ZeebeClock.newDefaultClock(container);
+      final Map<String, ZeebeClock> clocks = new HashMap<>();
+      cluster.getNodes().forEach((id, node) -> clocks.put(id, ZeebeClock.newDefaultClock(node)));
+      final Map<String, Instant> startTimes = new HashMap<>();
+      clocks.forEach((id, clock) -> startTimes.put(id, clock.getCurrentTime()));
 
       // when
-      final Instant startTime = clock.getCurrentTime();
       engine.increaseTime(offset);
-      final Instant endTime = clock.getCurrentTime();
 
       // then
-      assertThat(endTime).isCloseTo(startTime.plus(offset), within(10, ChronoUnit.SECONDS));
+      final Map<String, Instant> endTimes = new HashMap<>();
+      clocks.forEach((id, clock) -> endTimes.put(id, clock.getCurrentTime()));
+
+      assertThat(endTimes)
+          .allSatisfy(
+              (id, endTime) ->
+                  assertThat(endTime)
+                      .isCloseTo(startTimes.get(id).plus(offset), within(10, ChronoUnit.SECONDS)));
     }
   }
 }
